// libs/core/enums/dto/anime-type.dto.ts
/**
 * Acceptable values of anime types.
 */
export enum AnimeTypeDto {

  // Anime for TV.
  TV = 'TV',

  // Anime for promotional purposes.
  OVA = 'OVA',

  // Full-meter animation movie.
  MOVIE = 'MOVIE',

  // Anime episodes that aren't part of the original show.
  SPECIAL = 'SPECIAL',

  // Anime that is directly released onto the Internet.
  ONA = 'ONA',

  // Musical anime.
  MUSIC = 'MUSIC',
}